requirements.txt
python-telegram-bot>=20.7
firebase-admin>=6.4.0
python-dotenv>=1.0.0
httpx>=0.25.2
qrcode>=7.4.2
Pillow>=10.0.0
APScheduler>=3.10.4
flask>=3.0.0
gunicorn>=21.2.0
feedparser>=6.0.11
requests>=2.31.0
